source_code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
aiohappyeyeballs=2.4.4=py39haa95532_0
aiohttp=3.11.10=py39h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
alabaster=0.7.16=py39haa95532_0
arrow=1.3.0=py39haa95532_0
astroid=3.3.8=py39haa95532_0
asttokens=3.0.0=py39haa95532_0
async-timeout=5.0.1=py39haa95532_0
asyncssh=2.17.0=py39haa95532_0
atomicwrites=1.4.0=py_0
attrs=23.1.0=pypi_0
autobahn=23.6.2=pypi_0
automat=22.10.0=pypi_0
autopep8=2.0.4=pyhd3eb1b0_0
babel=2.16.0=py39haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
backports.tarfile=1.2.0=py39haa95532_0
bcrypt=4.3.0=py39h636fa0f_0
beautifulsoup4=4.13.4=py39haa95532_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=25.1.0=py39haa95532_0
blas=1.0=mkl
bleach=6.2.0=py39haa95532_0
bottleneck=1.4.2=py39hc99e966_0
brotlicffi=1.0.9.2=py39h5da7b33_1
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.8.3=pypi_0
cffi=1.15.1=pypi_0
chardet=4.0.0=py39haa95532_1003
charset-normalizer=3.1.0=pypi_0
click=8.1.8=py39haa95532_0
cloudpickle=3.1.1=py39haa95532_0
colorama=0.4.6=py39haa95532_0
comm=0.2.1=py39haa95532_0
constantly=15.1.0=pypi_0
cookiecutter=2.6.0=py39haa95532_1
cryptography=45.0.5=py39hcaec639_0
curl-cffi=0.12.0=pypi_0
debugpy=1.8.11=py39h5da7b33_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.13=py39haa95532_0
diff-match-patch=20200713=pyhd3eb1b0_0
dill=0.3.8=py39haa95532_0
docstring-to-markdown=0.11=py39haa95532_0
docutils=0.21.2=py39haa95532_0
exceptiongroup=1.2.0=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.7.1=h8ddb27b_0
flake8=7.1.1=py39haa95532_0
frozendict=2.4.6=pypi_0
frozenlist=1.5.0=py39h827c3e9_0
hyperlink=21.0.0=pypi_0
icu=73.1=h6c2663c_0
idna=3.4=pypi_0
imagesize=1.4.1=py39haa95532_0
importlib-metadata=8.5.0=py39haa95532_0
importlib_metadata=8.5.0=hd3eb1b0_0
incremental=22.10.0=pypi_0
inflection=0.5.1=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.29.5=py39haa95532_1
ipython=8.15.0=py39haa95532_0
isodate=0.6.1=pypi_0
isort=6.0.1=py39haa95532_0
jaraco.classes=3.2.1=pyhd3eb1b0_0
jaraco.context=6.0.0=py39haa95532_0
jaraco.functools=4.1.0=py39haa95532_0
jedi=0.19.2=py39haa95532_0
jellyfish=1.1.3=py39h8ecf97c_0
jinja2=3.1.6=py39haa95532_0
jpeg=9e=h827c3e9_3
jsonschema=4.25.0=py39haa95532_0
jsonschema-specifications=2023.7.1=py39haa95532_0
jupyter_client=8.6.3=py39haa95532_0
jupyter_core=5.8.1=py39haa95532_0
jupyterlab_pygments=0.3.0=py39haa95532_0
keyring=25.6.0=py39haa95532_0
krb5=1.20.1=h5b6d351_0
libclang=14.0.6=default_hb5a9fac_2
libclang13=14.0.6=default_h8e68704_2
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h70ee33d_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
logger=1.4=pypi_0
logzero=1.7.0=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markdown-it-py=2.2.0=py39haa95532_1
markupsafe=3.0.2=py39h827c3e9_0
matplotlib-inline=0.1.6=py39haa95532_0
mccabe=0.7.0=pyhd3eb1b0_0
mdurl=0.1.0=py39haa95532_0
mistune=3.1.2=py39haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h827c3e9_2
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
more-itertools=10.3.0=py39haa95532_0
multidict=6.1.0=py39h827c3e9_0
multitasking=0.0.12=pypi_0
mypy_extensions=1.0.0=py39haa95532_0
nbclient=0.10.2=py39haa95532_0
nbconvert=7.16.6=py39haa95532_0
nbconvert-core=7.16.6=py39haa95532_0
nbconvert-pandoc=7.16.6=py39haa95532_0
nbformat=5.10.4=py39haa95532_0
nest-asyncio=1.6.0=py39haa95532_0
numexpr=2.10.1=py39h4cd664f_0
numpy=2.0.1=py39h055cbcc_1
numpy-base=2.0.1=py39h65a83cf_1
numpydoc=1.9.0=py39haa95532_0
openssl=3.0.17=h35632f6_0
packaging=24.2=py39haa95532_0
pandas=2.3.1=py39h885b0b7_0
pandoc=2.12=haa95532_3
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.4=py39haa95532_0
pathspec=0.10.3=py39haa95532_0
peewee=3.18.2=pypi_0
pexpect=4.9.0=py39haa95532_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=25.1=pyhc872135_2
platformdirs=4.3.7=py39haa95532_0
pluggy=1.5.0=py39haa95532_0
ply=3.11=py39haa95532_0
prompt-toolkit=3.0.43=py39haa95532_0
propcache=0.3.1=py39h827c3e9_0
protobuf=6.31.1=pypi_0
psutil=5.9.0=py39h827c3e9_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycodestyle=2.12.1=py39haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.23.0=pypi_0
pydocstyle=6.3.0=py39haa95532_0
pyflakes=3.2.0=py39haa95532_0
pygithub=2.4.0=py39haa95532_0
pygments=2.19.1=py39haa95532_0
pyjwt=2.10.1=py39haa95532_0
pylint=3.3.5=py39haa95532_0
pylint-venv=3.0.3=py39haa95532_0
pyls-spyder=0.4.0=pyhd3eb1b0_0
pynacl=1.5.0=py39h7edc060_1
pyotp=2.8.0=pypi_0
pyparsing=3.1.0=pypi_0
pyqt=5.15.10=py39h5da7b33_1
pyqt5-sip=12.13.0=py39h827c3e9_1
pyqtwebengine=5.15.10=py39h5da7b33_1
pysocks=1.7.1=py39haa95532_0
python=3.9.23=h716150d_0
python-dateutil=2.8.2=pypi_0
python-fastjsonschema=2.20.0=py39haa95532_0
python-lsp-black=2.0.0=py39haa95532_1
python-lsp-jsonrpc=1.1.2=pyhd3eb1b0_0
python-lsp-server=1.12.2=py39hbc747e5_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2025.2=pyhd3eb1b0_0
pytoolconfig=1.2.6=py39haa95532_0
pytz=2023.3=pypi_0
pyuca=1.2=py39haa95532_1
pywin32=308=py39h5da7b33_0
pywin32-ctypes=0.2.2=py39haa95532_0
pyyaml=6.0.2=py39h827c3e9_0
pyzmq=26.2.0=py39h5da7b33_0
qdarkstyle=3.2.3=pyhd3eb1b0_0
qstylizer=0.2.2=py39haa95532_0
qt-main=5.15.2=h19c9488_12
qt-webengine=5.15.9=h5bd16bc_7
qtawesome=1.4.0=py39haa95532_0
qtconsole=5.6.1=py39haa95532_1
qtpy=2.4.1=py39haa95532_0
rdflib=6.3.2=pypi_0
rdflib-jsonld=0.6.2=pypi_0
referencing=0.30.2=py39haa95532_0
requests=2.31.0=pypi_0
rich=13.9.4=py39haa95532_0
rope=1.13.0=py39haa95532_0
rpds-py=0.22.3=py39h636fa0f_0
rtree=1.0.1=py39h2eaa2aa_0
setuptools=78.1.1=py39haa95532_0
simplejson=3.19.1=pypi_0
sip=6.7.12=py39h5da7b33_1
six=1.16.0=pypi_0
smartapi-python=1.4.8=pypi_0
snowballstemmer=2.2.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.5=py39haa95532_0
sphinx=7.3.7=py39h827c3e9_0
sphinxcontrib-applehelp=2.0.0=pyhd3eb1b0_1
sphinxcontrib-devhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.1.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=2.0.0=pyhd3eb1b0_1
sphinxcontrib-serializinghtml=2.0.0=pyhd3eb1b0_0
spyder=6.0.7=py39haa95532_1
spyder-kernels=3.0.5=py39hbc747e5_0
sqlite=3.50.2=hda9a48d_1
stack_data=0.2.0=pyhd3eb1b0_0
superqt=0.7.3=py39hbc747e5_0
tbb=2021.8.0=h59b6b97_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.6.3=py39hbc747e5_1
three-merge=0.1.1=pyhd3eb1b0_0
timedelta=2020.12.3=pypi_0
timestamp=0.0.1=pypi_0
tinycss2=1.4.0=py39haa95532_0
tk=8.6.14=h5e9d12e_1
tomli=2.0.1=py39haa95532_0
tomlkit=0.13.2=py39haa95532_0
tornado=6.5.1=py39h827c3e9_0
traitlets=5.14.3=py39haa95532_0
twisted=22.10.0=pypi_0
twisted-iocpsupport=1.0.4=pypi_0
txaio=23.1.1=pypi_0
typing-extensions=4.6.3=pypi_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
ujson=5.10.0=py39h5da7b33_1
unidecode=1.3.8=py39haa95532_0
urllib3=2.0.3=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
watchdog=4.0.2=py39haa95532_0
wcwidth=0.2.13=py39haa95532_0
webencodings=0.5.1=py39haa95532_1
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
whatthepatch=1.0.2=py39haa95532_0
wheel=0.45.1=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.17.0=py39h827c3e9_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
yapf=0.40.2=py39haa95532_0
yarl=1.18.0=py39h827c3e9_0
yfinance=0.2.65=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.21.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_1
zope-interface=6.0=pypi_0
zstd=1.5.6=h8880b57_0
